build: default - add line number hints to .po(t) files
Preparation for future default: omit line number hints, which reduces
churn in the repo when messages themselves haven't changed
(but line numbers they appear on have changed). This assists in CI so
code only changes won't make it appear as though i18n strings changed.

No tool in the openwrt build or translate process actively depends on
line numbers, but for the possibility of some po tools used locally.
Should you need line numbers in your *.po files, supply "--ln=full" to
this or the i18n-sync.sh script, which typically invokes this one.

To change the default, modify the line:

 my $add_location = "xxxx";

where "xxxx" is one of "never" "file" or "full"

Signed-off-by: Paul Donald <newtwen+github@gmail.com>