luci-mod-network: WiFi; mandate WPA3 and OWE security on 6GHz band
Reveal only permissible security on the 6GHz band (WiFi 6E, 7).

This should assist in deployments to avoid undefined or weird behaviour
when disallowed security are used in the 6GHz band. WiFi 6E and newer
mandate WPA3/SAE and OWE only. Some radios and firmware may not be
strict about this, but a number of client devices may be.

Signed-off-by: Paul Donald <newtwen+github@gmail.com>